+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - fix: --attach-key flag in CLI did not work
 - auto-import GPG key from file
 - auto-encrypt GPG possibility
+- S/MIME multiple recipients
 
 ## 0.9.9 (2020-02-10)
 - smime dependency is optional – thus package can be installed directly without having swig library